Add CfdHelper and increase compile sdk to 33
Summary:
CfdHelper is a singleton helper class that will act as the messenger between
our fragments and the CFD. It is setup by MainActivity. It will observe the
PresentationFragment's lifecycle and if a fragment wants something to be be
done on the CFD before PresentationFragment has been able to start up. Then
CfdHelper will queue the command and execute any queued commands in coroutines
after the PresentationFragment has reached it's Resumed state.

Test Plan:
The TransformFragment (the first fragment the user lands on) will attempt to
change the text on the CFD to say "Transform fragment wuz here" as soon as it
starts up. This will almost always happen too soon and be put in queue.
- Verify the CFD text reads "Trasnform fragment wuz here" on startup

To test that commands work without going through the queue tap on any image in
the TransformFragment.
- Verify the CFD text reads "Image at position <X> clicked"