Minor refactor to introduce an InfiniteSlidePager wrapper
Description:
Still working on figuring out a cleaner implementation for the infinite
slide item transitions, rather than just extending the contents of the
collection in perpetuity (which is horrible for memory, eventually).

This slight refactor should be beneficial if we need to tweak the `ViewPager2`
implementation, and helps move some boilerplate out of our `Fragment`.

Related Issue(s):
https://gpproduct.atlassian.net/browse/PPA-9661